rename customizer context to interface context
